Updated screenshot folder to place trades in foldes for the day and ticker
Updated error message to be No position found for instead of error getting (as this will happen when you don't have a position and is expected in this case). Also added curSymbol text to see if we have a 'symbol', but maybe a thread issue for getting the textbox value.

LastError is now readonly, and has scrollbars to help maybe see what is happening with some of the issues.